requirements.txt
Django>=5.2.9,<5.3
djangorestframework>=3.16.0,<4
psycopg2>=2.9.11,<2.10.0
drf-spectacular>=0.29.0,<0.30.0
django-cors-headers>=4.9.0,<5.0.0